Stop the container only if running

+++ scripts/before-install.sh
@@ -26,7 +26,9 @@ sudo service docker start
 fi
 
 echo 'Killing any container of the old Docker image'
-docker rm $(docker stop $(docker ps -a -q --filter ancestor=njetty/data_ingestor --format="{{.ID}}"))
+if [[ $(sudo docker ps -a -q --filter ancestor=njetty/data_ingestor --format="{{.ID}}") ]]; then
+	docker rm $(docker stop $(docker ps -a -q --filter ancestor=njetty/data_ingestor --format="{{.ID}}"))
+fi
 
 fuser -k 4000/tcp
